Handle edge cases: make distinction between cursor motion and typing
Previously, the plugin would move the markers no matter if the user had
actually inserted new lines or not.  This was mostly fine and
unnoticable, but in short files it would cause an error message when the
plugin tried to strip beyond the file's edge (it would wind up with a
range like -3,6 which is not allowed).

This version still does not handle the (admittedly obscure) input:

    i ^O O

(which inserts a new line while in insert mode), but this is arguably
not actually remaining in insert mode, and besides seems like a very
unusual thing to do.  If you are reading this and are having trouble
with that input or similar, let me know and I'll work something out.
Nobody has complained yet.