Clamp the bottom of the file in normal mode edits
This may have to do with the plugins I have installed, but other users
are getting an error when using undo/redo at the end of the file.  The
problem was that certain normal mode commands set the '] mark *past* the
end of the file for some reason.

Since I already clamp in insert mode, I see no reason not to do it here
as well.  I can't think of a situation where it would be incorrect.

Closes #4 on GitHub.